Add PSYLINK with patched fopen(, "r+b") for overlays
In MGS overlays there are many instances of uninitialized memory in
padding between globals in rdata and data sections. This padding is
generated by PSYLINK by fseeking forward. On modern systems this empty
space is filled with 0s, but it looks like on older systems sometimes
this space was filled with uninitialized memory.

In order to replicate this behavior, this commit patches PSYLINK to open
the resulting overlay file with fopen(, "r+b") instead of fopen(, "wb").
In such a configuration PSYLINK will write on top of existing data
in the overlay file. We will use it in mgs_reversing by creating a
dummy file filled with non-zero bytes and running the linker on top of
it. Those dummy bytes will be preserved in the resulting overlay and
we will be able to detect them and fill them with our own uninitialized
memory later.

PSYLINK has two functions related to fopen: FUN_004029EB (fopen
with "wb") and FUN_00402A35 (fopen with "r+b"). This commit patches
one of the calls to FUN_004029EB to instead call FUN_00402A35.

Patch contents:
$ diff <(objdump -S psylink.exe) <(objdump -S psylink_overlay_fopen_patch.exe)
2c2
< psylink.exe:     file format pei-i386
---
> psylink_overlay_fopen_patch.exe:     file format pei-i386
9482c9482
<   408a88:	e8 5e 9f ff ff       	call   0x4029eb
---
>   408a88:	e8 a8 9f ff ff       	call   0x402a35